+++ test/NosCore.Parser.Tests/CardParserTests.cs
@@ -25,7 +25,7 @@ namespace NosCore.Parser.Tests
     {
         private Mock<ILogLanguageLocalizer<LogLanguageKey>> _logLanguageMock = null!;
         private Mock<IDao<CardDto, short>> _cardDaoMock = null!;
-        private Mock<IDao<BCardDto, short>> _bCardDaoMock = null!;
+        private Mock<IDao<BCardDto, int>> _bCardDaoMock = null!;
         private string _tempFolder = null!;
         private List<CardDto> _savedCards = null!;
         private List<BCardDto> _savedBCards = null!;
@@ -35,7 +35,7 @@ namespace NosCore.Parser.Tests
         {
             _logLanguageMock = new Mock<ILogLanguageLocalizer<LogLanguageKey>>();
             _cardDaoMock = new Mock<IDao<CardDto, short>>();
-            _bCardDaoMock = new Mock<IDao<BCardDto, short>>();
+            _bCardDaoMock = new Mock<IDao<BCardDto, int>>();
             _savedCards = [];
             _savedBCards = [];
